requirements_nlp.txt
scikit-learn>=1.3.0
numpy<2.0.0,>=1.21.0
torch>=1.11.0
spacy>=3.0.0
optimum[onnxruntime]>=1.16.0
sentence-transformers>=2.2.2 